# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Digital ABD API
  version: 0.4.0
  description: |
    As-Built Documentation platform for Optical Fiber Cable (OFC) networks.

    **Authentication:** Bearer JWT (except `/health`, `/api/v1/auth/login`, `/api/v1/auth/oidc/*`).

    **Idempotency:** Include `Idempotency-Key` header on POST/PUT write operations.

    **Errors:** RFC 7807 Problem Details (`application/problem+json`).

    Live interactive docs available at `/docs` when the API server is running.
  contact:
    name: Digital ABD Program Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.digiabd.example.com
    description: Production (placeholder)
...
paths:
...
  /api/v1/projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      summary: Create project
      operationId: createProject
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
  /api/v1/projects/{projectId}/routes:
    get:
      tags: [Routes]
      summary: List routes for project
      operationId: listRoutes
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Route list
    post:
      tags: [Routes]
      summary: Create route
      operationId: createRoute
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteRequest'
      responses:
        '201':
          description: Route created
  /api/v1/projects/{projectId}/workflow/approval-chain:
    get:
      tags: [Deviations]
      summary: Get project deviation approval chain
      operationId: getProjectApprovalWorkflow
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: Approval chain
    put:
      tags: [Deviations]
      summary: Update project deviation approval chain
      operationId: updateProjectApprovalWorkflow
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approval_chain]
              properties:
                approval_chain:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Approval chain updated
  /api/v1/checklists/{projectType}:
    get:
      tags: [Projects]
      summary: Get mandatory checklist configuration by project type
      operationId: getChecklistConfig
      security: [{ bearerAuth: [] }]
      parameters:
        - name: projectType
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Checklist config
    put:
      tags: [Projects]
      summary: Update mandatory checklist configuration by project type
      operationId: updateChecklistConfig
      security: [{ bearerAuth: [] }]
      parameters:
        - name: projectType
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [required_items]
              properties:
                required_items:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Checklist config updated
...
  /api/v1/gis/wms/capabilities:
    get:
      tags: [GIS]
      summary: WMS service capabilities descriptor
      operationId: getWmsCapabilities
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: WMS capabilities
  /api/v1/gis/routes/{routeId}/geojson:
    get:
      tags: [GIS]
      summary: Export route as GeoJSON FeatureCollection
      operationId: exportGeoJson
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/routeId'
      responses:
        '200':
          description: GeoJSON FeatureCollection
  /api/v1/gis/routes/{routeId}/overlay:
    get:
      tags: [GIS]
      summary: Planned versus actual route overlay
      operationId: plannedActualOverlay
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/routeId'
      responses:
        '200':
          description: Overlay feature collection
  /api/v1/gis/routes/{routeId}/export:
    get:
      tags: [GIS]
      summary: Export route GIS artifact by format
      operationId: exportGisByFormat
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/routeId'
        - name: format
          in: query
          schema: { type: string, enum: [geojson, kml, shapefile] }
      responses:
        '200':
          description: GIS export artifact metadata
  /api/v1/cad/routes/{routeId}/generate:
    post:
      tags: [CAD]
      summary: Generate as-built CAD artifact
      operationId: generateCad
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/routeId'
      responses:
        '201':
          description: CAD artifact generated
  /api/v1/cad/routes/{routeId}/artifacts:
    get:
      tags: [CAD]
      summary: List CAD artifacts for route
      operationId: listCadArtifacts
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/routeId'
      responses:
        '200':
          description: CAD artifact list
  /api/v1/etl/routes/{routeId}/jobs:
    post:
      tags: [ETL]
      summary: Queue ETL job for route
      operationId: queueEtlJob
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/routeId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                job_type:
                  type: string
                  enum: [gis_layer_refresh, cad_generation, data_quality_scan]
      responses:
        '201':
          description: Job queued
  /api/v1/etl/jobs:
    get:
      tags: [ETL]
      summary: List ETL jobs
      operationId: listEtlJobs
      security: [{ bearerAuth: [] }]
      parameters:
        - name: route_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: ETL job list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from /api/v1/auth/login or OIDC flow
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
    routeId:
      name: routeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string }
        vendor_name: { type: string, nullable: true }
        client_name: { type: string, nullable: true }
        project_type: { type: string, enum: [urban, rural, hdd_heavy, mixed] }
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        client_name: { type: string }
        vendor_name: { type: string }
        project_type: { type: string, enum: [urban, rural, hdd_heavy, mixed] }
    CreateRouteRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        total_length_km: { type: number }
...